+++ Modules/Product/Http/Controllers/Admin/ProductQuestionController.php
@@ -11,20 +11,22 @@ use Modules\Product\Http\Requests\Admin\ProductQuestionUpdateRequest;
 use Modules\Product\Http\Requests\Admin\ProductQuestionCreateRequest;
 use Modules\Product\Http\Resources\ProductQuestionResource;
 use Modules\Product\Models\ProductQuestion;
+use Modules\Product\Repositories\ProductQuestionRepository;
 use Modules\Product\Services\Qna\ProductQuestionStorage;
 
 class ProductQuestionController extends Controller
 {
     public function __construct(
-        protected ProductQuestionStorage $productQuestionStorage
-    ) {
-        $this->authorizeResource(ProductQuestion::class, 'product_question');
-    }
+        protected ProductQuestionStorage $productQuestionStorage,
+        protected ProductQuestionRepository $productQuestionRepository,
+    ) {}
 
     public function store(
         ProductQuestionCreateRequest $request,
     ): ProductQuestionResource
     {
+        $this->authorize('create', ProductQuestion::class);
+
         $productQuestion = $this->productQuestionStorage->store(ProductQuestionDto::fromFormRequest($request));
 
         return new ProductQuestionResource($productQuestion);
@@ -32,9 +34,13 @@ class ProductQuestionController extends Controller
 
     public function update(
         ProductQuestionUpdateRequest $request,
-        ProductQuestion $product_question
+        int $product_question
     ): ProductQuestionResource
     {
+        $product_question = $this->productQuestionRepository($product_question);
+
+        $this->authorize('update', $product_question);
+
         $product_question = $this->productQuestionStorage->update(
             $product_question,
             ProductQuestionDto::fromFormRequest($request)
@@ -43,8 +49,12 @@ class ProductQuestionController extends Controller
         return new ProductQuestionResource($product_question);
     }
 
-    public function destroy(ProductQuestion $product_question): Response
+    public function destroy(int $product_question): Response
     {
+        $product_question = $this->productQuestionRepository($product_question);
+
+        $this->authorize('delete', $product_question);
+
         $this->productQuestionStorage->delete($product_question);
 
         return \response()->noContent();
@@ -52,9 +62,13 @@ class ProductQuestionController extends Controller
 
     public function approve(
         ProductQuestionApproveOrRejectRequest $request,
-        ProductQuestion $product_question
+        int $product_question
     ): Response
     {
+        $product_question = $this->productQuestionRepository($product_question);
+
+        $this->authorize('approve', $product_question);
+
         $this->productQuestionStorage->changeStatus(
             $product_question,
             ProductQuestionStatus::fromValue(ProductQuestionStatus::APPROVED)
@@ -70,9 +84,13 @@ class ProductQuestionController extends Controller
 
     public function reject(
         ProductQuestionApproveOrRejectRequest $request,
-        ProductQuestion $product_question
+        int $product_question
     ): Response
     {
+        $product_question = $this->productQuestionRepository($product_question);
+
+        $this->authorize('reject', $product_question);
+
         $this->productQuestionStorage->changeStatus(
             $product_question,
             ProductQuestionStatus::fromValue(ProductQuestionStatus::REJECTED)
